avoid new type constraints from TypeScript 2.x

+++ jsweet-transpiler/src/test/java/source/init/ArrayNew.java
@@ -27,26 +27,26 @@ public class ArrayNew {
 		strings1 = new String[] { "a", "b", "c" };
 		assert strings1.length == 3;
 		array(strings1).push("d");
-		assert strings1.length == 4;
+		assert (int)strings1.length == 4;
 		assert strings1[3] == "d";
 
 		Array<String> strings3 = array(new String[] { "a", "b", "c" });
 		assert strings3.length == 3;
 		strings3.push("d");
-		assert strings3.length == 4;
+		assert (int)strings3.length == 4;
 		assert strings3.$get(3) == "d";
 		
 		strings3 = new Array<String>("a", "b", "c" );
 		assert strings3.length == 3;
 		strings3.push("d");
-		assert strings3.length == 4;
+		assert (int)strings3.length == 4;
 		assert strings3.$get(3) == "d";
 
 		strings3 = new Array<String>();
 		strings3.push("a", "b", "c");
 		assert strings3.length == 3;
 		strings3.push("d");
-		assert strings3.length == 4;
+		assert (int)strings3.length == 4;
 		assert strings3.$get(3) == "d";
 		
 	}
